Converted UUID generation from v4 to v1.
V4 generators require a good source or randomness and the sparseness of the likely coverage of the set of all possible UUIDs to guarantee uniqueness of UUIDs to some very high likelihood.  V1 generators rely on the promise of uniqueness of MAC addresses and an upper bound on the rate of UUID generation to guarantee uniqueness.

In our case, our hardware doesn't come with a good source of randomness, we won't be having the database create millions of UUIDs per second.  Also, there are no security reasons for us not expose the MAC address where a UUID was generated or the time it was generated.  These reasons make the V1 generator a better choice for us.