CORE-5602 Fixed template avu data ID validation.
Added a valid-uuid-to-path function to donkey.services.filesystem.metadata-templates that will return the path for the given data UUID, or throw a not-found exception.
Used this function in the metadata template avu pre-hook calls since path-exists is not null-safe, and now the valid-uuid-to-path will return an error message referring to the invalid UUID instead of referring to a null path.