CORE-5797 Updated template avu endpoints.
Updated the donkey.persistence.metadata functions for template AVUs to set and update the created_by and modified_by columns instead of the owner_id column.
Updated the Metadata Template AVU CRUD endpoints to no longer use the owner_id field, to no longer return the "user" field in JSON responses, and to check for write permissions for the create, update, and delete endpoints.