CORE-5832 Refactored app rating endpoints.
Renamed /secured/rate-analysis and /secured/delete-rating to POST and DELETE /apps/:id/rating endpoints in both Donkey and metadactyl.
Added swagger docs to metadactyl endpoints.
Updated kameleon.queries/get-existing-user-id so that it's a public function.
CORE-2675 Updated the responses of these endpoints to include the total number of ratings on the app.