CORE-5832 Fix for update/create Pipeline endpoints.
Refactored the korma.db/transaction call out of the metadactyl.zoidberg/add-pipeline and update-pipeline functions and into the functions they call.
There seems to be an issue where exceptions are not rolling back the sql queries inside the transactions in these function which are called by the compojure-api routes. Moving the transaction calls inside the add-pipeline-app and update-pipeline-app functions seems to do the trick.