CORE-5832 Refactor Pipeline JSON; added /apps/:id/tasks
Refactored /apps/:id/file-parameters into a /apps/:id/tasks endpoint that returns an App in a format similar to Pipeline JSON, but without 'steps' and 'mapping' fields.
Refactored Pipeline JSON to match this format, since there was always only 1 app in the 'apps' array, which was the pipeline itself, so those fields have been moved up out of the obsolete 'apps' array.
Refactored the add and update pipeline endpoints to return the full pipeline JSON, instead of an array of 1 app ID.