CORE-6155 Updated parameter_groups name and label.
The UI is only submitting labels for parameter groups, and if more than 1 group is submitted, the "name" field is omitted in all but the first group.
Since the UI has been using parameter group "label" fields instead of the "name" fields for a long time, we switched their NOT NULL constraints in the db.
Updated the app model schema in metadactyl to require a group's label.