CORE-6201 Fixed metadactyl pipeline submission JSON.
The metadactyl.analyses.base buildStep methods did not have access to the outputs of previous steps, so pipeline input params after the first step were always empty. Passing an empty string to me.raynes.fs/base-name returns either an empty string or the current working directory if running metadactyl from the command-line.
Updated the preprocessor function in the metadactyl.analyses.base/DeJobRequestFormatter buildInputArgs method to return nil if it gets a blank input param value.
Updated metadactyl.analyses.params/remove-irods-home to also return nil if it gets a blank path.
Fixed the logic in metadactyl.analyses.params/value-for-param to use the right io-maps value, rather than the input param ID key, to get the correct output-value-map value.
Updated metadactyl.analyses.common/load-steps to use a reduce instead of a mapv with the buildStep function, and updated the buildStep and buildConfig functions in metadactyl.analyses.common and metadactyl.analyses.base to accept all previous steps in addition to the current step. This allows the metadactyl.analyses.base/buildConfig methods to pass all outputs from the current and previous steps to the buildParams methods.