CORE-6506 Fix for /apps/:id/is-publishable
Updated the query in metadactyl.app-validation/task-ids-for-app to select only from the app_steps table.
Selecting from the apps table and joining on the app_steps table still returned a result for an app with no steps, which caused this function to return a non-empty list (with only 1 nil element), and this caused the is-publishable endpoint to incorrectly return true for apps with no steps.