CORE-6670 Refactored data-info /data/directory/create endpoint.
Renamed data-info /data/directory/create endpoint to /data/directories and updated it to accept an array of paths instead of only 1 path.
Refactored the data-info.services.create/create function to create multiple paths, but to only set own permissions on the top-most created directories, since the 'set-owner' function is already recursive.
Added a /secured/filesystem/directories endpoint to Donkey that is just a passthrough to the data-info endpoint.
Refactored the /secured/filesystem/directory/create endpoint to send a request to the data-info /data/directories endpoint and respond with the stat info of the new directory.