Refactor data-info.services.exists
Refactored `with-jargon` and `user-exists` check outside of loop that calls `path-exists-for-user?` and into `do-exists`.
Used `juxt` instead of `hash-map`.
No longer need to URL decode the POST body of the request, which can actually prevent a path like "/testZone/home/test/foo%20bar%20baz" from being found, which is a valid path.