CORE-6827 Migrated metadata db calls out of data-info.
The metadata save endpoint in the data-info service now sends a request to the metadata service to list metadata AVUs, instead of connecting to the metadata db directly.
Refactored metadata db connection setup and AVU queries out of kameleon and into the metadata service.
Updated compojure-api and related lib versions in data-info.