Refactor data-info client in Donkey
This moves the actual HTTP calls to the service into a new donkey.clients.data-info.raw namespace, which is a basic/dumb passthrough. The pre-existing donkey.clients.data-info now uses this, isolating extra processing (and pesky things potentially still using Jargon or other bits of Donkey) in that namespace. This namespace-shifting will also make it easier to migrate jargon calls to use data-info endpoints.

Calls made outside of donkey.clients.data-info are not yet moved here, though that is intended to happen.